requirements.txt
allure-pytest==2.9.45
allure-python-commons==2.9.45
astor==0.8.1
atomicwrites==1.4.0
attrs==21.3.0
bleach==4.1.0
certifi==2021.10.8
charset-normalizer==2.0.9
colorama==0.4.4
comtypes==1.1.10
configparser==5.2.0
crayons==0.4.0
cx-Oracle==8.3.0
docutils==0.18.1
et-xmlfile==1.1.0
execnet==1.9.0
flake8==4.0.1
future==0.18.2
greenlet==1.1.2
helium==3.0.8
idna==3.3
importlib-metadata==4.8.2
iniconfig==1.1.1
keyring==23.4.0
mccabe==0.6.1
openpyxl==3.0.9
packaging==21.3
parse==1.19.0
parse-type==0.5.2
pkginfo==1.8.2
pluggy==1.0.0
psutil==5.8.0
psycopg2-binary==2.9.2
py==1.11.0
pycodestyle==2.8.0
pyflakes==2.4.0
Pygments==2.10.0
pyparsing==3.0.6
pytest==6.2.5
pytest-env==0.6.2
pytest-failed-screenshot==1.0.2
pytest-forked==1.3.0
pytest-rerunfailures==10.2
pytest-xdist==2.4.0
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2021.3
pywin32-ctypes==0.2.0
readme-renderer==32.0
requests==2.26.0
requests-toolbelt==0.9.1
responses==0.10.14
rfc3986==1.5.0
selene==2.0.0a40
selenium==3.141.0
six==1.16.0
SQLAlchemy==1.4.29
testrail-api==1.10.3
toml==0.10.2
tqdm==4.62.3
typing_extensions==4.0.1
urllib3==1.26.7
webdriver-manager==3.5.0
webencodings==0.5.1
zipp==3.6.0
